feat: add uploadFile function

+++ src/services/upload.service.js
@@ -1,12 +1,23 @@
 const { s3Client } = require("../config/aws.config");
+const fs = require("fs");
 
-const handleFileSend = async (body, key) => {
-  const res = await s3Client.putObject({
+const uploadFile = async (filepath, key) => {
+  if (!fs.existsSync(filepath)) return;
+
+  const fileStream = fs.createReadStream(filepath);
+
+  fileStream.on("error", function (err) {
+    throw err;
+  });
+
+  await s3Client.putObject({
     Bucket: process.env.BUCKET_SPACE,
     ACL: "public-read",
-    Body: body,
-    Key: key,
+    Body: fileStream,
+    Key: `profiles/${key}`,
   });
+
+  return `${BUCKET_URL}/profiles/${key}`;
 };
 
-module.exports = { handleFileSend };
+module.exports = { uploadFile };
